Add readme and heroku trash

+++ src/main/java/com/senderman/durkafeedbackbot/service/StoryPublisherService.java
@@ -3,6 +3,7 @@ package com.senderman.durkafeedbackbot.service;
 import com.annimon.tgbotsmodule.api.methods.Methods;
 import com.annimon.tgbotsmodule.services.CommonAbsSender;
 import com.senderman.durkafeedbackbot.dbservice.FeedbackService;
+import com.senderman.durkafeedbackbot.model.Feedback;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -17,12 +18,15 @@ public class StoryPublisherService {
         this.durkaChannelId = durkaChannelId;
     }
 
-    public void publish(int id, long issueChatId, CommonAbsSender sender) throws StoryNotFoundException {
+    public void publish(int id, long issueChatId, CommonAbsSender sender) throws StoryNotFoundException, NotAStoryException {
         var feedbackOptional = feedbackService.findById(id);
         if (feedbackOptional.isEmpty())
             throw new StoryNotFoundException("Not found feedback with id " + id);
 
         var feedback = feedbackOptional.get();
+        if (!feedback.getType().equals(Feedback.TYPE_TRUE_STORY)) {
+            throw new NotAStoryException("Feedback with id " + id + "is not a story!");
+        }
         var text = "<b>История от подписчика</b>\n\n" + feedback.getMessage();
         feedbackService.deleteById(id);
 
@@ -35,6 +39,12 @@ public class StoryPublisherService {
 
     public static class StoryNotFoundException extends Exception {
         public StoryNotFoundException(String message) {
+            super(message);
+        }
+    }
+
+    public static class NotAStoryException extends Exception {
+        public NotAStoryException(String message) {
             super(message);
         }
     }
